refactor: Overall Overhaul
I think given the difficulty of making this work, trying to accommmodate
a GUI (as an inexperienced GUI programmer) was really getting in the way.

So I've started again as a CLI. With the idea of wrapping it all in
a slick GUI somewhere in the distant future.

This is still a work in progress.
So far the groundwork for rendering the master file and sidecar data
is in place, as well as some loose functionality on rendering segments
and splicing the master file.

It's exciting progress though!